Initial Code Migration
This project was developed during a Red Hat Consulting engagement.  The system was deployed into multiple production environments.  The code has been stripped of all identifying marks and released to the community as ODIE - OpenShift
Disconnected Installer for Enterprise.

Sadly, the version history was lost during this migration.  The original system was developed by contributions from the following developers:

- Mike Battles
- Kevin Franklin
- Nate Revo
- Randy Johnson
- Stuart Bain
- Stephen Palmer
- Whitney Champion
- Glenn Snead
- Paul Begovich